Add item stat rollover when player walks over collectible object
A UI element called UiItemDetail has been added. This element is shown/hidden when the player walks over a Collectible item. Upon contact, the item's stats (stored in a SO) are sent to the UI to be displayed. This feature has been put on the bronze coin at the start.